🚚 refactor: move SessionList
