+++ src/pages/chat/features/SessionList/CollapseGroup.tsx
